navMesh error FIXED
My best guesses for this would be to double check when you're building
that you have the right scene selected in the 'Scenes In Build' section
of the Build Settings window. There's two scenes in the project by
default (the starting point and the completed scene version), so it may
be picking up the wrong one.